[libpas] MTE-tag WebContent.EnhancedSecurity processes https://bugs.webkit.org/show_bug.cgi?id=300638 rdar://162538653
Reviewed by Mark Lam.

Such processes are considered 'hardened' a la Lockdown Mode processes,
at least with regards to MTE tagging and the enablement of MTE hardening
features like adjacent-tag-exclusion.
This patch ensures that they are treated as such, including full tagging
of all non-compact medium-and-small objects allocated through libpas.

Specifically, the processes which this will affect are:
  - com.apple.WebKit.WebContent.EnhancedSecurity
  - com.apple.WebKit.WebContent.EnhancedSecurity.Development

We need to switch to getprogname() instead of pbi_name/pbi_comm because
pbi_name is size-limited to the first 32B of the process name.

Canonical link: https://commits.webkit.org/301970@main