[WebXR] Test gardening after 300383@main https://bugs.webkit.org/show_bug.cgi?id=301241 rdar://161189145
Reviewed by Mike Wyrzykowski.

The change in 300383@main is trying to be "a good citizen" by
unregistering observing visibility changes from WebXRSession::~WebXRSession.
This is causing an exception due to trying to accessing already
destructed hash map during DOMWindow destruction.

Since the observers are kept in a WeakHashSet there is no need to
unregister during WebXRSession::~WebXRSession.

In addtion, when obtaining permissions for an inline session, the
request is forwarded to PlatformXRCompositor, which only handles
immersiveVR sessions and asserts in Debug builds because the methods
are called in the wrong order.

Instead of forwarding the request to the compositor for inline mode,
all permissions are granted unconditionally.

* Source/WebCore/Modules/webxr/WebXRSession.cpp:
(WebCore::WebXRSession::~WebXRSession):
* Source/WebKit/UIProcess/XR/PlatformXRSystem.cpp:
(WebKit::PlatformXRSystem::requestPermissionOnSessionFeatures):

Canonical link: https://commits.webkit.org/301975@main