[Layout] Fix crash in WebCore::RenderGrid::gridAreaRangeForOutOfFlow https://bugs.webkit.org/show_bug.cgi?id=300354 rdar://161639703
Reviewed by Alan Baradlay.

Fix a crash in WebCore::RenderGrid::gridAreaRangeForOutOfFlow where we haven't completed
in-flow placement and grid sizing yet, and so our positions vector doesn't have valid values.

This aligns us closer with the spec, where only in flow children are
eligible for margin collapsing.

The root cause of this crash is we're attempting to do margin collapsing with an out of flow child, but
we can only do this for in flow children.

[1] <https://drafts.csswg.org/css2/#collapsing-margins>

* LayoutTests/fast/css-grid-layout/grid-crash-area-range-for-out-of-flow-expected.txt: Added.
* LayoutTests/fast/css-grid-layout/grid-crash-area-range-for-out-of-flow-no-text-expected.txt: Added.
* LayoutTests/fast/css-grid-layout/grid-crash-area-range-for-out-of-flow-no-text.html: Added.
* LayoutTests/fast/css-grid-layout/grid-crash-area-range-for-out-of-flow.html: Added.
* Source/WebCore/rendering/RenderBlock.cpp:
(WebCore::RenderBlock::isSelfCollapsingBlock const):

Canonical link: https://commits.webkit.org/301976@main