[ macOS Tahoe wk2 Debug x86_64 ] ipc/serialized-type-info.html is a constant text failure https://bugs.webkit.org/show_bug.cgi?id=301021 rdar://162900827
Reviewed by Abrar Rahman Protyasha.

This was caused by https://commits.webkit.org/301207@main and appears
to be fixed by https://commits.webkit.org/301947@main. The test expectations
added in https://commits.webkit.org/301750@main as a part of this bug have
not been removed, so to close this bug, we do that in this patch.

* LayoutTests/platform/mac-wk2/TestExpectations:

Canonical link: https://commits.webkit.org/302036@main