[CSS Zoom] Fix calc(em) for unzoomed properties https://bugs.webkit.org/show_bug.cgi?id=301351 rdar://163267333
Reviewed by Sammy Gill.

When canonicalizing to "px" we currently try to compensate
the fact that the `computeNonCalcLengthDouble` would use the
compute font size (that has zoom applied to it) by dividing
by the zoom factor here. This should no longer be necessary
for unzoomed properties (when the evaluation flag is on) as
we now resolve it against the unzoomed computed font size.

Once all properties are converted to be unzoomed and the flag
is removed we can remove the counter branch.

* LayoutTests/fast/css/line-height-basics.html:
* Source/WebCore/css/CSSToLengthConversionData.cpp:
(WebCore::CSSToLengthConversionData::evaluationTimeZoomEnabled const):
* Source/WebCore/css/CSSToLengthConversionData.h:
* Source/WebCore/css/calc/CSSCalcTree+Simplification.cpp:
(WebCore::CSSCalc::canonicalize):

Canonical link: https://commits.webkit.org/302041@main