[GTK][WPE] Crash in JSWebXRGLBinding_createProjectionLayer https://bugs.webkit.org/show_bug.cgi?id=301341
Reviewed by Fujii Hironori.

In 560926ef@main we enabled the WebXR Layers switch by default for
GTK and WPE ports. That's a mistake because the API is not ready
yet to be used. That was making many WebXR sessions crash because
they're trying to use the layers API when it isn't ready yet.

No new tests required, as we're just disabling a runtime switch by
default. We had to skip two previously unskipped tests as they
require the switch to be on to pass.

* LayoutTests/platform/glib/TestExpectations:
* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:

Canonical link: https://commits.webkit.org/302046@main