[LCP] Optimize computing LCP entries for text https://bugs.webkit.org/show_bug.cgi?id=301370 rdar://163285757
Reviewed by Tim Nguyen.

LCP rects for text are unioned for all the text nodes that contribute to an Element,
so we can't do early area checks. However, when we get to `generateLargestContentfulPaintEntry()`
we've collected all the rects, so now we can do an area check (if there are no transforms in the
ancestor chain).

* Source/WebCore/page/LargestContentfulPaintData.cpp:
(WebCore::LargestContentfulPaintData::generateLargestContentfulPaintEntry):
(WebCore::LargestContentfulPaintData::didLoadImage):

Canonical link: https://commits.webkit.org/302072@main