[visionOS] Exiting Docked mode for YouTube videos shows black screen for a sec. https://bugs.webkit.org/show_bug.cgi?id=301397 rdar://163076430
Reviewed by Jer Noble.

297873@main introduced a regression causing the video frame to temporaly flicker
while the content is being repainted.
In addition, following 301011@main an additional black frame can be noticed.
This occurs because the teardown of the AVSampleBufferVideoRenderer occurs
much more quickly that it used to (the transition between an AVSBDL to a figVideoTarget
is now very fast as we no longer have to seek).
We now wait an extra 1s, before tearing down the AVSBVR containing the figVideoTarget).
This removes the black frames.
Ideally we should enqueue frames to both AVSBVR and AVSBDL until a new frame
is actually visible. We will implement this solution in a follow-up change.
For now, we keep the change small and simple to facilitate integration.

Manually tested, the current testing infrastructure doesn't allow for docking the video.

* Source/WebCore/platform/graphics/avfoundation/AudioVideoRendererAVFObjC.mm:
(WebCore::AudioVideoRendererAVFObjC::stageVideoRenderer):

Canonical link: https://commits.webkit.org/302104@main