Increase HTMLNameCache max capacity to reduce cache misses https://bugs.webkit.org/show_bug.cgi?id=300227 rdar://162019828
Reviewed by Ryosuke Niwa.

We currently use the HTMLNameCache to cache frequently seen attribute names and values in the HTMLFastPathParser.
HTMLNameCache uses two caches under the hood - an AtomString cache for storing attributes and a QualifiedName cache for storing common attribute names. They both have the same max size.

The maximum cache size was initially set to 512 for both. Increasing the max size to 2048 for AtomString cache has these effects:

On SP3, it increases the hit rate from from 93.35% -> 96.71% (+3.36%) on SP3.
On a real-world workload: it increases hit rate from 56.11% -> 60.61% (Real-world: +4.5%)

This is mostly insiginficant overall on SP3, but it is a .75 - 1.2%% improvement on TODOMVC-Jquery subtest.

* Source/WebCore/html/parser/HTMLNameCache.h:

Canonical link: https://commits.webkit.org/302119@main