[Perf] Anly check if the decompression session accepts a format description when the format changes rdar://163408718 https://bugs.webkit.org/show_bug.cgi?id=301467
Reviewed by Jean-Yves Avenard.

Performance testing shows an observable CPU increase in the GPU and videocodecd processes as
a result of calling VTDecompressionSessionCanAcceptFormatDescription(). This is needed in order
to determine whether the decompression session needs to be torn down and re-built when the
video stream's format changes. However, it doesn't need to be called for every frame; its safe
to assume that the decompression session can continue to decode the same format.

* Source/WebCore/platform/graphics/cocoa/WebCoreDecompressionSession.h:
* Source/WebCore/platform/graphics/cocoa/WebCoreDecompressionSession.mm:
(WebCore::WebCoreDecompressionSession::ensureDecompressionSessionForSample):

Canonical link: https://commits.webkit.org/302151@main