Array Allocation Sinking Should Track Initialized Indicies https://bugs.webkit.org/show_bug.cgi?id=301468 rdar://162617198
Reviewed by Yusuke Suzuki.

After fixing conditional writes in 300888@main there was still a remaining issue.
If there ended up being a read from a conditionally initialized index of the
array we could end up forwarding the hole value to that read rather than `undefined`.

To fix this we add a bit vector that tracks whether a particular index is guaranteed
to be initialized on every path to this point.

Lastly this change removes an unused size argument passed to the ArrayButterly's
Allocation. As of this change that would trigger an additional allocation and is
unused.

Tests: JSTests/stress/array-sink-diamond-initialization-then-read.js
       JSTests/stress/array-sink-read-uninitialized-hole.js
       JSTests/stress/array-sink-conditional-initialization.js

Canonical link: https://commits.webkit.org/302153@main