REGRESSION (iOS 26): Sliding side menu looks bad (fixed element not extending under address bar) https://bugs.webkit.org/show_bug.cgi?id=301425 rdar://163352309
Reviewed by Abrar Rahman Protyasha.

Adjust the fixed color extension heuristic to err on the side of extending colors past the edges of
the layout viewport in the case where the hit-tested element is not viewport-sized on the sampled
edge, but _is_ viewport-sized or larger on the adjacent side. In practice, this causes us to
automatically apply color extensions in cases where the fixed element is similar to a sidebar that
fills up at least half of the width of the viewport, instead of leaving gaps either above or below
the sidebar.

Test: fast/page-color-sampling/color-sampling-sidebar-and-header.html

* LayoutTests/fast/page-color-sampling/color-sampling-sidebar-and-header-expected.txt: Added.
* LayoutTests/fast/page-color-sampling/color-sampling-sidebar-and-header.html: Added.
* Source/WebCore/page/LocalFrameView.cpp:
(WebCore::LocalFrameView::fixedContainerEdges const):

Also give this scenario its own candidate result type (`IsSidebar`), so that we can defer to any
existing sampled fixed colors on that edge to maintain color sampling stability (similar to what we
do for popups and other viewport-sized overlays).

Canonical link: https://commits.webkit.org/302171@main