Add a debug assertion for subclasses of CanMakeCheckedPtr to be always heap allocated https://bugs.webkit.org/show_bug.cgi?id=301458
Reviewed by Geoffrey Garen.

This PR adds a debug assertion in the destructor of CanMakeCheckedPtrBase to assert that
the deletion is initiated with operator delete of the corresponding class.

Since there are quite a few cases where CheckedPtr capable type is either a member variable
of another class or it's allocated in stack, this PR also adds the mechanism to suppress
the newly added debug assertion in CanMakeCheckedPtrBase, and deploys it in places that
currently do not use operator delete for destruction.

Finally, this PR also converts a few objects to be heap-allocated on its own instead of as
a member variable of another type or in stack.

No new tests since there should be no behavioral differences.

Canonical link: https://commits.webkit.org/302184@main