[GTK][WPE] Get render process hardware acceleration information for webkit://gpu from the web process https://bugs.webkit.org/show_bug.cgi?id=301423
Reviewed by Nikolas Zimmermann.

We try to create a PlatformDisplay in the UI process with the same
parameters as the process to get the hardware acceleration information.
We could get the information from the web process using an IPC message
to make sure the data we get is accurate and to allow getting
information for other data we don't currently show because it's not
available in the UI process.

Canonical link: https://commits.webkit.org/302185@main