[JSC] Fix assert in array.new_elem slow path https://bugs.webkit.org/show_bug.cgi?id=301466 rdar://163407647
Reviewed by Yusuke Suzuki.

There are 3 elem segment types: passive, active, or declarative. Only passive
elem segments are available at runtime and get a non-nullptr `element`.
array.new_elem can still refer to active and declarative elem segments, usually
this traps, because those element segments would be considered 0-lengthed and
the array being created usually have length > 0. This doesn't trap when the
array size is 0, but in that case the `element` is a nullptr and the debug
asserts crash. This PR fixes the asserts to allow the array to be 0-lengthed.

New test cases are added to array_new_elem.js

Canonical link: https://commits.webkit.org/302224@main