Crash in WebCore::JPEGXLImageDecoder::decode https://bugs.webkit.org/show_bug.cgi?id=301514
Reviewed by Carlos Garcia Campos.

JPEGXLDecoder::decode assumes that if JPEGXLDecoder::processInput fails,
it will return JXL_DEC_ERROR. This is a reasonable assumption, but there
is one case where it does not hold. Fix it.

Also, unskip the layout test that's designed to catch precisely this
problem.

Canonical link: https://commits.webkit.org/302235@main