Disable nullability completeness https://bugs.webkit.org/show_bug.cgi?id=301411 rdar://163319634
Reviewed by Geoffrey Garen.

This disables the clang nullability completeness warning. Swift/C++ interop
requires us to label some APIs as _Nullable or _Nonnull, and this warning then
requires us to similarly annotate all the other APIs in the same file. This
leads to annotation proliferation, and when we're not careful this can even
become viral across multiple files if we expand macros to include these
annotations. This is not currently deemed helpful or beneficial for WebKit.
See Slack discussion at

https://webkit.slack.com/archives/CU64U6FDW/p1761239537764009
Canonical link: https://commits.webkit.org/302237@main