Negative hover selector not working properly on hover in lockdown mode https://bugs.webkit.org/show_bug.cgi?id=301090 rdar://163481884
Reviewed by Alan Baradlay.

We fail to invalidate for :not(:hover) when the CSS selector JIT is disabled.

Test: fast/css/hover-invalidation-no-jit.html
* LayoutTests/fast/css/hover-invalidation-no-jit-expected.html: Added.
* LayoutTests/fast/css/hover-invalidation-no-jit.html: Added.
* Source/WebCore/css/SelectorChecker.cpp:
(WebCore::SelectorChecker::checkOne const):

Remove the special case for invalidation mode. It is wrong in negation context and our current invalidation system does not need it.
There is no corresponding code in SelectorCompiler anymore.

Canonical link: https://commits.webkit.org/302245@main