[Layout] Harden RenderGrid::gridAreaRangeForOutOfFlow() code for scenarios we haven't completed layout https://bugs.webkit.org/show_bug.cgi?id=300934 rdar://162810289
Reviewed by Sammy Gill.

In some situations, we can get into a state where we are attempting to access the `positions` vector in
but haven't completed in-flow placement and grid sizing yet.

We should add an assert to verify we're not doing anything improperly previously in the code, and return something basic
that doesnt access unbuilt data structures on production builds.

* Source/WebCore/rendering/RenderGrid.cpp:
(WebCore::RenderGrid::gridAreaRangeForOutOfFlow const):

Canonical link: https://commits.webkit.org/302263@main