Change fast/css/viewport-units-zoom.html to use setPageZoomFactor https://bugs.webkit.org/show_bug.cgi?id=301404 rdar://163316825
Reviewed by Lily Spiniolas.

This test was added in 196138@main in order to address a bug related to
page scale factor and viewport relative units. It seems like viewport
relative units should *not* be affected by page scale factor. However,
the test uses css-zoom to try and replicate this behavior which is
actually supposted to scale viewport units.

To address this we can change the test to use the setPageZoomFactor
in Internals. The test itself has a div with fixed height and width that
uses viewport units and sets the page scale factor. The expected is
basically the same with the height multiplied by 2. When the page scale
factor is applied, the width of the div should remain the same but the
height should get doubled like in the expectation.

* LayoutTests/fast/css/viewport-units-zoom-expected.html:
* LayoutTests/fast/css/viewport-units-zoom.html:

Canonical link: https://commits.webkit.org/302275@main